split query constructor utilities into smaller files
